Adds BLE error codes
In case that BLE is turned of or if Location is not enabled, an exception is thrown which is
currently not handled. This commit simply adds the corresponding error codes as a comment.

If an unknown error is thrown, this should probably be shown to the user in a dialog.